backend/requirements.txt
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.2
django-filter==2.0.0
django-oauth-toolkit==1.2.0
djangorestframework==3.9.0
idna==2.7
Markdown==3.0.1
oauthlib==3.0.0
psycopg2==2.7.5
psycopg2-binary==2.7.7
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.5
pyyaml==3.12
requests==2.20.1
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.24.1
python-dateutil==2.8.0
